+++ .github/dependabot.yaml
@@ -1,7 +1,7 @@
 version: 2
 updates:
   - package-ecosystem: "npm"
-    directory: "/src/"
+    directory: "/apps/generator"
     target-branch: "develop"
     schedule:
       interval: "monthly"
