+++ apps/README.md
@@ -1,0 +1,6 @@
+# apps
+
+- generator
+  - カレンダーを生成するプログラム。
+  - `dist` 内に生成したカレンダーを `docs/dist` にコピーして公開する仕組み。
+    - `docs`を GitHubPages で公開しています。
