+++ apps/generator/master/playables.yaml
