+++ apps/generator/src/consts.ts
