+++ apps/generator/src/models/CalendarEvent.ts
