+++ apps/generator/src/models/Playable.ts
