+++ apps/generator/tsconfig.json
@@ -9,6 +9,6 @@
     "skipLibCheck": true,
     "forceConsistentCasingInFileNames": true
   },
-  "include": ["ts/**/*.ts"],
+  "include": ["src/**/*.ts"],
   "types": ["@types/jest"]
 }
