--- src/tsconfig.eslint.json
@@ -1,5 +1,0 @@
-// ESLint用に別ファイルで定義
-{
-  "extends": "./tsconfig.json",
-  "include": ["ts/**/*", "__tests__/**/*", "jest.config.ts", "*.config.ts"]
-}
